Compute loc for libarchive input

+++ src/srcml/parse_request.hpp
@@ -44,7 +44,7 @@ struct ParseRequest {
     boost::optional<std::string> disk_filename;
     int position;
     int status;
-    int loc;
+    long loc;
 };
 
 #endif
